I don't think everyone has cmake, if this is a bad idea, feel free to revert.
SVN revision: 7500

+++ legacy/ecore/conf-e-x11.sh
@@ -1,8 +1,13 @@
 #!/bin/sh
 
-cmake clean distclean
-rm config.cache
-./configure \
+if [ -f ./Makefile ]; then
+    make clean distclean
+fi
+if [ -f ./config.cache ]; then
+    rm config.cache
+fi
+
+./autogen.sh \
 --enable-ecore-x \
 --enable-ecore-evas \
 --enable-ecore-evas-gl \
